XMonad: Display workspaces on polybar via FIFO
This way, we can mark what workspaces are visible on other screens,
which was previously impossible. It should also allow us to use the
RenameWorkspaces package as we are now in control of exactly how the
workspace names are displayed.

On the down side, it loses us the ability to click to change workspaces,
but this *should* be possible to reimplement using polybar escape
sequences + xmonad messaging